% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_points_pattern.R
\name{create_points_pattern}
\alias{create_points_pattern}
\title{Generate points proportionally to surface chlorophyll-a values.}
\usage{
create_points_pattern(raster, multi = 1, mld = 0)
}
\arguments{
\item{raster}{input chlorophyll RasterLayer}

\item{multi}{multiplicator to control number of points}

\item{mld}{mixed layer depth}
}
\value{
list with matrix of longitude, latitude, mixed layer depth
}
\description{
Transform surface-chlorophyll values into a randomly distributed relative number of points on a plane depending on the value of each cell.
A multiplicator value is used to control the total number of points. A mixed layer depth is defined in the third column (particles are at the surface at zero, below surface at negative values)
}
\examples{
 
data(surface_chl)
pts <- create_points_pattern(surface_chl, multi=1000)
plot(surface_chl)
points(pts,cex=0.1)

}

