% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractROMS.r
\name{extractROMS}
\alias{extractROMS}
\title{Extract single ROMS time step.}
\usage{
extractROMS(roms, time = 1)
}
\arguments{
\item{roms}{ROMS object}

\item{time}{index into 4th dim of i_u,i_v,i_w in ROMS object}
}
\value{
list single time slice ROMS
}
\description{
See \link{toyROMS} for an example.
}

