% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopit_V1.R
\name{loopit}
\alias{loopit}
\title{Loopit}
\usage{
loopit(speed, runtime, trajectories = FALSE)
}
\arguments{
\item{speed}{(w_sink) sinking rate m/days}

\item{runtime}{(time) total number fo days to run the model}

\item{trajectories}{TRUE/FALSE statement to define whether to store particle trajectories (model runs much faster without storing trajectories). Default is FALSE.}
}
\value{
list(pts=pts, pend=pend, stopindex=obj$stopindex, ptrack=obj$ptrack, lon_list=lon_list, idx_list=idx_list, idx_list_2D=idx_list_2D, id_list=id_list)
}
\description{
Loop trackit_3D to follow particles through different consecutive ROMS-sclices. Looping can also increase performance when using very large number of particles by looping through shorter time steps.
Loops are set to run in half day intervals. If no runtime is defined, the function will loop depending on the depth of the deepest cell and the sinking speed to allow each particle to possibly sink to the seafloor (2*max(h)/speed)
}
\examples{
data(surface_chl)
data(toyROMS)
pts <- create_points_pattern(surface_chl, multi=100)
h <- toyROMS$h
all_i_u <- toyROMS$i_u
all_i_v <- toyROMS$i_v
all_i_w <- toyROMS$i_w


track <- loopit(speed = 100, runtime = 50, romsobject = toyROMS)

plot(pts)
points(track$pnow, col="red")

library(rgl)
plot3d(pts, zlim = c(-1500,1))
plot3d(track$pnow, col="red", add=TRUE)

}

