% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopit_trackit_2D.R
\name{loopit_trackit_2D}
\alias{loopit_trackit_2D}
\title{Loopit_Trackit 2D}
\usage{
loopit_trackit_2D(pts, romsobject, w_sink = 100, time = 50)
}
\arguments{
\item{pts}{input points}

\item{w_sink}{sinking rate m/days}

\item{time}{total number of days to run the model}

\item{kdtree}{kd tree}
}
\value{
list(ptrack = ptrack, pnow = pnow, plast = plast, stopindex = stopindex, indices = indices, indices_2D = indices_2D)
}
\description{
Function called by loopit to track particles through a 2D-ROMS-field.
}

