% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptrackr-package.r
\docType{package}
\name{ptrackr}
\alias{ptrackr}
\alias{ptrackr-package}
\title{ptrackr - a package to track individual particles through 2D- and 3D-space}
\description{
The main function is loopit_2D3D which calls other functions from the package to do the tracking.
More details in the individual desription files...
}

