% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_knn.R
\name{setup_knn}
\alias{setup_knn}
\title{Setup knn lookup to which ROMS-cell a particle belongs}
\usage{
setup_knn(lon_roms = lon_u, lat_roms = lat_u, depth_roms = hh)
}
\arguments{
\item{lon_roms}{longitude values of ROMS grid}

\item{lat_roms}{latitude values of ROMS grid}

\item{depth_roms}{depth values of ROMS grid}
}
\value{
list, kdtree, kdxy
}
\description{
Setup k-nearest-neighbour upfront for use in the particle tracking loops.
}

