% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildparams.r
\name{buildparams}
\alias{buildparams}
\title{buildparams}
\usage{
buildparams(speed, time_step_in_s, p0 = 1030, p1 = 1100, cosO = 1,
  g = 9.81, K = 0.0025, E = 1, r = 0.00016, Ucsq = -(0.05 * (p0 - p1)
  * g * 2 * (1.5 * E)^(1/3) * r)/(p0 * K), Wd = speed,
  SedFunct = function(U_div, dens) time_step_in_s * (p1 * (dens) * Wd * cos0 *
  (U_div) * (U_div))/p0)
}
\arguments{
\item{speed}{sinking speed}

\item{time_step_in_s}{the time-step in seconds}

\item{r}{particle radius}
}
\value{
list(p0 = p0, p1 = p1, cosO = cosO, g = g, K = K, E = E, r = r, Wd = Wd, Ucsq = Ucsq, SedFunct = SedFunct)
}
\description{
Function to set up parameters for the sedimentation equation used in the trackit_2D function
}

