% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_points_pattern.R
\name{create_points_pattern}
\alias{create_points_pattern}
\title{Generate points proportionally to surface chlorophyll-a values.}
\usage{
create_points_pattern(a_chl_raster, a_multiplicator = 1,
  a_mixed_layer_depth = 0)
}
\arguments{
\item{a_chl_raster}{input chlorophyll RasterLayer}

\item{a_multiplicator}{scale}

\item{a_mixed_layer_depth}{mixed layer depth}
}
\value{
list with matrix of longitude, latitude, mixed layer depth
}
\description{
Transform surface-chlorophyll values into a randomly distributed relative number of points depending on the value of each cell
needs a chlorophyll raster and a multiplicator value to determine the number of points
mixed layer depth can be used to assign the third dimension (particles start from the surface at 0, from below surface at negative values)
}

