% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_package_ptrack_trackit_3D.R
\name{trackit_3D}
\alias{trackit_3D}
\title{Trackit 3D}
\usage{
trackit_3D(pts, romsobject, w_sink = 86.4, time = 50)
}
\arguments{
\item{pts}{input points}

\item{w_sink}{sinking rate m/days, time is days}

\item{time}{days}

\item{kdtree}{kd tree}
}
\value{
list(ptrack = ptrack, pnow = pnow, plast = plast, stopindex = stopindex, indices = indices, indices_2D = indices_2D)
}
\description{
Function to track particles through a ROMS-field.
}
\details{
the function needs an input for speed of the sinking particles (w_sink) and for time
due to the limitation of RAM available, time is restricted depending on the number of particles
(too long runs might give an error because the generated vector is too large)
I found half days work great
}

